Added getxmp() for TIFF
