+++ Tests/test_file_tiff.py
@@ -600,6 +600,16 @@ class TestFileTiff:
         with Image.open(outfile) as reloaded:
             assert "icc_profile" not in reloaded.info
 
+    def test_xmp(self):
+        with Image.open("Tests/images/lab.tif") as im:
+            xmp = im.getxmp()
+
+            assert isinstance(xmp, dict)
+
+            description = xmp["xmpmeta"]["RDF"]["Description"]
+            assert description[0]["format"] == "image/tiff"
+            assert description[3]["BitsPerSample"]["Seq"]["li"] == ["8", "8", "8"]
+
     def test_close_on_load_exclusive(self, tmp_path):
         # similar to test_fd_leak, but runs on unixlike os
         tmpfile = str(tmp_path / "temp.tif")
